* dev: (28 commits)
  colors final (hopefully) touches
  background color
  Update SunPower endpoint response
  org colors fixed
  user color fixes
  data not found screen
  default cfv set fns to override false
  fix calendar colors
  more color fixes
  scroll some tables to top on page change
  fixed header for contact, user, org
  add state to orgs list
  Installation Agreements: Query update
  Installation Agreements: Query update
  changed query stuff
  fix dates on project screen
  fix date selections for round robin
  remove column from company dash
  SMS Inbox: Fixed footer, fix for page numbers
  tweaks and fixes
  ...

+++ src/main/resources/db/migration/procedures/brs/commissions/R__get_all_users_overrides_earned_in_payroll_snapshot.sql
@@ -29,7 +29,7 @@ BEGIN
 
     case when v_position_id = 1 then
         RETURN QUERY select results.closer,
-                            results.project_id,
+                            results.project_id::bigint,
                             results.customer_name,
                             results.system_size,
                             results.overrides_earned,
@@ -81,7 +81,7 @@ BEGIN
                      where results.overrides_earned - results.prior_pay != 0;
         when v_position_id = 4 then
             RETURN QUERY select results.closer,
-                                results.project_id,
+                                results.project_id::bigint,
                                 results.customer_name,
                                 results.system_size,
                                 results.overrides_earned,
