+++ src/main/resources/db/migration/procedures/flow/search/R__search_projects_with_down_line.sql
@@ -66,9 +66,9 @@ BEGIN
   case
     when p_searchterm is not null and p_searchterm != '' then
       RETURN QUERY
-      SELECT limited_projects.id,
+      SELECT limited_projects.id::bigint,
              limited_projects.project_name,
-             limited_projects.contact_id,
+             limited_projects.contact_id::bigint,
              limited_projects.date_created,
              limited_projects.street1,
              limited_projects.street2,
@@ -78,13 +78,13 @@ BEGIN
              limited_projects."postalCode",
              limited_projects.latitude,
              limited_projects.longitude,
-             limited_projects.company_project_status_type_id,
+             limited_projects.company_project_status_type_id::bigint,
              limited_projects.project_status_type,
              limited_projects.contact
       FROM (select *
-            from (select p.id,
+            from (select p.id::bigint,
                          p.project_name,
-                         p.contact_id,
+                         p.contact_id::bigint,
                          p.date_created,
                          p.street1,
                          p.street2,
@@ -94,7 +94,7 @@ BEGIN
                          p.postal_code                            as "postalCode",
                          p.latitude,
                          p.longitude,
-                         p.company_project_status_type_id,
+                         p.company_project_status_type_id::bigint,
                          cpst.project_status_type,
                          (select row_to_json(contact1)
                           from (select c.id,
@@ -118,9 +118,9 @@ BEGIN
                             cpst.id = p_company_project_status_type_id
                           else 1 = 1 end
                   union
-                  select p.id,
+                  select p.id::bigint,
                          p.project_name,
-                         p.contact_id,
+                         p.contact_id::bigint,
                          p.date_created,
                          p.street1,
                          p.street2,
@@ -130,7 +130,7 @@ BEGIN
                          p.postal_code                            as "postalCode",
                          p.latitude,
                          p.longitude,
-                         p.company_project_status_type_id,
+                         p.company_project_status_type_id::bigint,
                          cpst.project_status_type,
                          (select row_to_json(contact1)
                           from (select c.id,
@@ -169,9 +169,9 @@ BEGIN
             limit p_limit offset p_offset) as limited_projects;
     else
       return query
-      SELECT limited_projects.id,
+      SELECT limited_projects.id::bigint,
                 limited_projects.project_name,
-                limited_projects.contact_id,
+                limited_projects.contact_id::bigint,
                 limited_projects.date_created,
                 limited_projects.street1,
                 limited_projects.street2,
@@ -181,12 +181,12 @@ BEGIN
                 limited_projects."postalCode",
                 limited_projects.latitude,
                 limited_projects.longitude,
-                limited_projects.company_project_status_type_id,
+                limited_projects.company_project_status_type_id::bigint,
                 limited_projects.project_status_type,
                 limited_projects.contact
-         FROM (select p.id,
+         FROM (select p.id::bigint,
                       p.project_name,
-                      p.contact_id,
+                      p.contact_id::bigint,
                       p.date_created,
                       p.street1,
                       p.street2,
@@ -196,7 +196,7 @@ BEGIN
                       p.postal_code                            as "postalCode",
                       p.latitude,
                       p.longitude,
-                      p.company_project_status_type_id,
+                      p.company_project_status_type_id::bigint,
                       cpst.project_status_type,
                       (select row_to_json(contact1)
                        from (select c.id,
